logR_connect {logR}	R Documentation

Connect logR to postgres

Description

Wrapper on DBI::dbConnect with defaults from envir.

Usage

logR_connect(host = Sys.getenv("POSTGRES_HOST", "127.0.0.1"),
  port = Sys.getenv("POSTGRES_PORT", "5432"),
  dbname = Sys.getenv("POSTGRES_DB", "postgres"),
  user = Sys.getenv("POSTGRES_USER", "postgres"),
  password = Sys.getenv("POSTGRES_PASSWORD", "postgres"), quoted = FALSE)

Arguments

host
character hostname/ip by default from ENV var 'POSTGRES_HOST'

port
character port by default from ENV var 'POSTGRES_PORT'

dbname
character port by default from ENV var 'POSTGRES_DB'

user
character port by default from ENV var 'POSTGRES_USER'

password
character port by default from ENV var 'POSTGRES_PASSWORD'

quoted
logical if TRUE it will return unevaluated expression which can be passed to remote R session. Database authentication will be included in expression.

Value

logical, if successfully connected then TRUE, otherwise FALSE. Use getOption("logR.conn") to access actual connection, which is set as side effect.

[Package logR version 2.1.3 Index]