logR_query {logR}	R Documentation

Query logR data

Description

Query logR logs from defined database or csv file.

Usage

logR_query(alert, status, since, where, limit, .conn = getOption("logR.conn"),
  .table = getOption("logR.table"), .schema = getOption("logR.schema"))

Arguments

alert
logical, if TRUE then filter out non-alerts.

status
logical, if TRUE then filter out success.

since
POSIXt or Date, if provided then filter on start timestamp.

where
list named, each element can be a vector of character values for 'IN' filter for name in list.

limit
integer, if provided then limit numbers of rows returned.

.conn
DBI connection.

.table
character.

.schema
character.

Details

By default all function arguments will be taken from options which are used to setup logR logging, for arguments description see logR.

Value

Logs according to filters in DESC order.

See Also

logR, logR_watcher

[Package logR version 2.1.3 Index]