logR_schema {logR}	R Documentation

Populate logR schema

Description

Function will execute scripts create schema and table to store logs. To view scripts see schema_sql.

Usage

logR_schema(meta = list(), drop = FALSE, silent = TRUE,
  .conn = getOption("logR.conn"))

Arguments

meta
list of metadata columns, postgres data types as character string named by column name like list(tag='VARCHAR(255)', val_col='DOUBLE PRECISION', int_col='INTEGER', systimecol='TIMESTAMPTZ'). Avoid reserved postgres keywords.

drop
logical, drop before creation.

silent
logical, silently try calls to db

.conn
DBI connection.

See Also

schema_sql, logR

[Package logR version 2.1.5 Index]