logR_watcher {logR}	R Documentation

Detects invalid logs

Description

On unlikely fatal error R session crash while evaluting logged expression it will not update status field in database. This functions helps to detect such cases.

Usage

logR_watcher(since = Sys.time() - 86400)

Arguments

since
POSIXt or Date, must be provided for this function.

Details

since parameter default value is current timestamp minus one day, so it should be scheduled daily. If you don't want to provide since param then use logR_query directly.

Value

Will raise warning in case if it will find any NULL stats logs. Otherwise NULL invisibly.

See Also

logR_query

[Package logR version 2.1.3 Index]