sql_val {logR}	R Documentation

Prepare SQL values

Description

Wraps character types into single quotes, numbers as non scientific.

Usage

sql_val(col, x, trunc_char_n = 252L)

Arguments

col
character vector of column names.

x
data.table a one row data.table.

trunc_char_n
integer number of characters to truncate character after. Default 252, ... is added to the end of string so the result will have 255 characters.

Value

character vector

[Package logR version 2.1.3 Index]